backend/requirements.txt

Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Cors
python-dotenv
pymysql
